+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### `FlexibleSearch` enhancements
 - Introduced action header for `.fxs` files [#547](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/547)
 - Added actions for connection settings in the header [#549](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/549)
+- Added `Execute` action to the header [#552](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/552)
 
 ### `ImpEx` enhancements
 - Introduced action header for `.impex` files [#550](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/550)